refactor(python): split framework.go by Django/annotation/FastAPI concern
framework.go (553 lines) held three concerns, not one Django axis, so split it
into django.go (ORM relations + URL routing), annotations.go (type-annotation
composes, shared by Django/Pydantic/FastAPI), and a slim framework.go (FastAPI
routes/Depends + decorator dispatch + shared arg/attr helpers).

- Decompose the four over-threshold functions into named helpers, all under
  the gate: emitSingleURLPattern -> emitURLViewCall/emitAsViewEdge/
  emitURLViewRef; emitTypeAnnotationEdge -> emitIdentifier/Generic/
  AttributeAnnotation; emitDjangoModelField -> djangoFieldName/
  emitModelRelationTarget; emitHarvest -> harvestMentions/Dispatch/PythonReach.
- Add a real re_path attribute-view fixture.

Behavior unchanged: the 14 python goldens (Django/FastAPI/Flask/SQLAlchemy
included) are byte-identical, and a digest over 1,842 Python stdlib files is
byte-identical between the pre- and post-split binaries.
